+++ app/dashboard/components/feed-goals.tsx
@@ -1,6 +1,6 @@
 import { getGoals } from "../actions";
-import GoalCard from "./GoalCard";
-import ClientPaginationControls from "./ClientPaginationControl";
+import GoalCard from "./goal-card";
+import ClientPaginationControls from "./client-pagination-controls";
 
 interface GoalsProps {
     page: number;
